Add blobstore/put-if-absent.
I decided that I did not want higher levels to have to keep track of
what seems to exist in the store. It would eliminate extra round-trips
(and perhaps costs) in the case of duplicate blocks, but it introduces
significant complexity and difficulty testing properly. I'd rather
have the backend take the hit of the extra round-trip.